Validation Token

+++ src/main/java/com/Project/Project/app/UserController.java
@@ -2,8 +2,7 @@ package com.Project.Project.app;
 
 import com.Project.Project.Business.UserService;
 import com.Project.Project.DataAcess.User;
-import com.Project.Project.ResponseHandler.DTO.TokenDTO;
-import com.Project.Project.ResponseHandler.DTO.UserDTO;
+import com.Project.Project.DTO.UserLoginDTO;
 import com.Project.Project.ResponseHandler.ResponseJSONhandler;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,12 +22,15 @@ public class UserController {
     private UserService userService;
     @PostMapping
     @RequestMapping("/user/cadastro")
-    public ResponseEntity cadastro(@RequestBody @Valid User user) {
-        String email = user.getEmail();
-        String nome = user.getName();
-        String cpf = user.getCpf();
-        String senha = user.getSenha();
-        return userService.verifyEmailCpfNameSenha(email, cpf, nome, senha, user);
+    public ResponseEntity cadastro(@RequestBody @Valid User user, @RequestHeader(HttpHeaders.AUTHORIZATION)String tkn) {
+        if(userService.tokenValdition(tkn)){
+
+            if(userService.tokenTypeUser(tkn).equalsIgnoreCase("cliente") || userService.tokenTypeUser(tkn).equalsIgnoreCase("fornecedor")){
+                throw new RuntimeException("Acesso Invalido");
+            }
+            return userService.verifyEmailCpfNameSenha(user.getEmail(), user.getCpf(), user.getName(), user.getSenha(), user);
+        }
+        throw new RuntimeException("Token invalido ");
     }
 
     @GetMapping
@@ -37,7 +39,7 @@ public class UserController {
     public List<User> ListUser(@RequestHeader(HttpHeaders.AUTHORIZATION)String tkn){
         if(userService.tokenValdition(tkn)){
 
-                if(userService.tokenTypeUser(tkn).equalsIgnoreCase("cliente")){
+                if(userService.tokenTypeUser(tkn).equalsIgnoreCase("cliente") || userService.tokenTypeUser(tkn).equalsIgnoreCase("fornecedor")){
                     throw new RuntimeException("Acesso Invalido");
                 }
 
@@ -49,35 +51,59 @@ public class UserController {
     @GetMapping
     @RequestMapping("/user/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public User FindId(@PathVariable Long id){
-        return userService.ListUniqClient(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+    public User FindId(@PathVariable Long id, @RequestHeader(HttpHeaders.AUTHORIZATION)String tkn){
+        if(userService.tokenValdition(tkn)){
+            if(userService.tokenTypeUser(tkn).equalsIgnoreCase("cliente") || userService.tokenTypeUser(tkn).equalsIgnoreCase("fornecedor")){
+                throw new RuntimeException("Acesso Invalido");
+            }
+            return userService.ListUniqClient(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "user not found"));
+        }
+        throw new RuntimeException("Token invalido ");
     }
 
     @GetMapping
     @RequestMapping("/user/login")
-    public TokenDTO loginUser(@RequestBody UserDTO userDTO ){
-         return userService.loginClient(userDTO);
+    public UserLoginDTO loginUser(@RequestBody UserLoginDTO userLoginDTO ){
+         return userService.loginClient(userLoginDTO);
     }
     @PutMapping("/user/update/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public User Update(@PathVariable("id") Long id, @RequestBody User user) {
-        String pass;
-        PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-        pass = passwordEncoder.encode(user.getSenha());
-        user.setSenha(pass);
-        return userService.UpdateById(id, user);
+    public User Update(@PathVariable("id") Long id, @RequestBody User user, @RequestHeader(HttpHeaders.AUTHORIZATION)String tkn) {
+        if(userService.tokenValdition(tkn)){
+            if(userService.tokenTypeUser(tkn).equalsIgnoreCase("cliente") || userService.tokenTypeUser(tkn).equalsIgnoreCase("fornecedor")){
+                throw new RuntimeException("Acesso Invalido");
+            }
+            String pass;
+            PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+            pass = passwordEncoder.encode(user.getSenha());
+            user.setSenha(pass);
+            return userService.UpdateById(id, user);
+        }
+        throw new RuntimeException("Token invalido ");
     }
     @PatchMapping("/user/updateEmail/{id}/{email}")
-    public User UpdateEmail(@PathVariable("id") Long id, @PathVariable("email") String email){
-        User user = FindId(id);
-        user.setEmail(email);
-        return userService.save(user);
+    public User UpdateEmail(@PathVariable("id") Long id, @PathVariable("email") String email, @RequestHeader(HttpHeaders.AUTHORIZATION)String tkn){
+        if(userService.tokenValdition(tkn)){
+            if(userService.tokenTypeUser(tkn).equalsIgnoreCase("cliente") || userService.tokenTypeUser(tkn).equalsIgnoreCase("fornecedor")){
+                throw new RuntimeException("Acesso Invalido");
+            }
+            User user = FindId(id, null);
+            user.setEmail(email);
+            return userService.save(user);
+        }
+        throw new RuntimeException("Token invalido ");
     }
     @DeleteMapping("/user/delete/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseJSONhandler delete (@PathVariable Long id){
-        FindId(id);
-        return userService.delete(id);
+    public ResponseJSONhandler delete (@PathVariable Long id, @RequestHeader(HttpHeaders.AUTHORIZATION)String tkn){
+        if(userService.tokenValdition(tkn)){
+            if(userService.tokenTypeUser(tkn).equalsIgnoreCase("cliente") || userService.tokenTypeUser(tkn).equalsIgnoreCase("fornecedor")){
+                throw new RuntimeException("Acesso Invalido");
+            }
+            FindId(id, null);
+            return userService.delete(id);
+        }
+        throw new RuntimeException("Token invalido ");
     }
